+++ src/contents/ui/configGeneral.qml
@@ -19,42 +19,47 @@ import "../js/layouts.js" as Layouts
 Kirigami.FormLayout {
 	Layout.fillWidth: true
 
-	property alias cfg_layoutKey: layoutKey.text
+	property alias cfg_layoutKey: layoutSelector.selectedLayoutKey
 	property alias cfg_useUserLayout: useUserLayout.checked
 	property alias cfg_useSpecificLocaleEnabled: useSpecificLocaleEnabled.checked
 	property alias cfg_useSpecificLocaleLocaleName: useSpecificLocaleLocaleName.text
+	property alias cfg_useCustomFont: useCustomFont.checked
+	property alias cfg_customFont: fontSelector.selectedFont
 
-	Text {
-		visible: false
-		id: layoutKey
-	}
-
-	PlasmaComponents.ComboBox {
-		Kirigami.FormData.label: i18n('Layout')
+	LayoutSelector {
+		id: layoutSelector
 		enabled: !cfg_useUserLayout
-		textRole: "text"
-		model: []
-		Component.onCompleted: {
-			// populate model from Theme object
-			var tmp = []
-			var idx = 0
-			var currentIdx = undefined
-			for(const key in Layouts.layouts) {
-				var name = Layouts.layouts[key]['name']
-				tmp.push({'value':key, 'text': name})
-				if (key === plasmoid.configuration['layoutKey']) currentIdx = idx
-				idx++
-			}
-			model = tmp
-
-			currentIndex = currentIdx
-		}
-		onCurrentIndexChanged: cfg_layoutKey = model[currentIndex]['value']
-	} // ComboBox
+		Kirigami.FormData.label: i18n('Layout')
+	}
 
 	PlasmaComponents.CheckBox {
 		id: useUserLayout
 		text: i18n("Use user layout")
+	}
+
+	Item {
+		height: 10
+	}
+
+	PlasmaComponents.CheckBox {
+		id: useCustomFont
+		text: i18n("Use custom font")
+	}
+	RowLayout {
+		enabled: cfg_useCustomFont
+
+		ColumnLayout {
+			PlasmaComponents.Label {
+				text: i18n('Font: %1', cfg_customFont.family)
+			}
+			PlasmaComponents.Label {
+				text: i18n('Size: %1', cfg_customFont.pointSize)
+			}
+		}
+
+		ConfigFontSelector {
+			id: fontSelector
+		}
 	}
 
 	Item {